+++ src/main/java/frc/robot/commands/pneumatics/arm/Arm.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.pneumatics.tilter;
+package frc.robot.commands.pneumatics.arm;
 
 import edu.wpi.first.wpilibj.*;
 import edu.wpi.first.wpilibj.command.Command;
